Change CBLDatabase.count type to uint64_t
* Change CBLDatabase.count type to uint64_t
* Implmented Database.count to Swift

+++ Objective-C/CBLDatabase.h
@@ -90,7 +90,7 @@ typedef struct {
 @property (readonly, nonatomic, nullable) NSString* path;
 
 /** The number of documents in the database. */
-@property (readonly, nonatomic) NSUInteger count;
+@property (readonly, nonatomic) uint64_t count;
 
 /** The database's configuration. If the configuration is not specify when initializing 
     the database, the default configuration will be returned. */
